Time: 0 ms (100%), Space: 94.9 MB (93.63%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -66,7 +66,7 @@ public:
             return (double)(ind1el);
         }
 
-        return double((double)(ind1el+ ind2el))/2.0;
+        return (ind1el+ ind2el)/2.0;
         
 
         //-----------------------------------Optimal----------------------------------------------
